reorganize class files
Created a new folder for the data classes and moved "DataAccessor" and "MemoryDataAccessor" to the new "data" folder.  Ensured program still compiled and ran.
